requirement.txt
PySide6>=6.6
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.23
websockets>=12.0
python-dotenv>=1.0.0
google-genai>=1.59.0
psutil>=5.9.0